requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
MarkupSafe==2.1.1
passlib==1.7.4
Pillow==9.2.0
PyMySQL==1.0.2
python-decouple==3.6
python-dotenv==0.21.0
sqlparse==0.4.3
Werkzeug==2.2.2
django-settings-export==1.2.1
pytest-django==4.5.2
factory-boy==3.2.1
